--- uniffi/src/message/mod.rs
@@ -1,13 +1,0 @@
-mod from_prior;
-mod message;
-mod pack_encrypted;
-mod pack_plaintext;
-mod pack_signed;
-mod unpack;
-
-pub use from_prior::{FromPriorExt, OnFromPriorPackResult, OnFromPriorUnpackResult};
-pub use message::MessageExt;
-pub use pack_encrypted::OnPackEncryptedResult;
-pub use pack_plaintext::OnPackPlaintextResult;
-pub use pack_signed::OnPackSignedResult;
-pub use unpack::OnUnpackResult;
