+++ uniffi/src/test_vectors/secrets/alice.rs
@@ -1,7 +1,7 @@
 use lazy_static::lazy_static;
 use serde_json::json;
 
-use didcomm::secrets::{Secret, SecretMaterial, SecretType};
+use didcomm_core::secrets::{Secret, SecretMaterial, SecretType};
 
 lazy_static! {
     pub(crate) static ref ALICE_SECRET_AUTH_KEY_ED25519: Secret = Secret {
